🚚 move: MVC 폴더 구조로 변경
