+++ src/model/Discount.js
